# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API заявок на расчёт межпланетного перелёта: характеристическая скорость (Δv), масса топлива, энергия.",
        "title": "Interplanetary Flight API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/interplanetaryflightrequests": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interplanetaryflightrequests"
                ],
                "summary": "Список заявок (создатель — свои; модератор — все)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Статус",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "formed_at с (RFC3339)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "formed_at по (RFC3339)",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "items",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/interplanetaryflightrequests/cart-icon": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interplanetaryflightrequests"
                ],
                "summary": "Иконка корзины (черновик)",
                "responses": {
                    "200": {
                        "description": "id, count",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/interplanetaryflightrequests/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interplanetaryflightrequests"
                ],
                "summary": "Детальная заявка с расчётами Δv / топливо / энергия",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.InterplanetaryFlightRequestDetail"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interplanetaryflightrequests"
                ],
                "summary": "Редактирование черновика заявки",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Масса сухого аппарата, Isp",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateInterplanetaryFlightRequestBody"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interplanetaryflightrequests"
                ],
                "summary": "Удалить черновик заявки",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/interplanetaryflightrequests/{id}/form": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interplanetaryflightrequests"
                ],
                "summary": "Сформировать заявку (draft → formed)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.InterplanetaryFlightRequestDetail"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/interplanetaryflightrequests/{id}/moderate": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interplanetaryflightrequests"
                ],
                "summary": "Модерация заявки (только модератор)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "complete | reject",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.ModerateInterplanetaryFlightRequestBody"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.InterplanetaryFlightRequestDetail": {
            "type": "object",
            "properties": {
                "completed_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "engine_isp_sec": {
                    "type": "number"
                },
                "formed_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.InterplanetaryFlightRequestLineFlatDTO"
                    }
                },
                "spacecraft_dry_mass_kg": {
                    "type": "number"
                },
                "status": {
                    "type": "string"
                },
                "total_fuel_mass_kg": {
                    "type": "number"
                }
            }
        },
        "handler.InterplanetaryFlightRequestLineFlatDTO": {
            "type": "object",
            "properties": {
                "delta_v_ms": {
                    "description": "Текущий расчёт по формулам лаб. 2 (удобно и для черновика)",
                    "type": "number"
                },
                "energy_j": {
                    "type": "number"
                },
                "from_orbit_radius_km": {
                    "type": "number"
                },
                "image_key": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "interplanetary_flight_description": {
                    "type": "string"
                },
                "interplanetary_flight_from": {
                    "type": "string"
                },
                "interplanetary_flight_title": {
                    "description": "Поля межпланетного перелёта (услуги) в том же объекте",
                    "type": "string"
                },
                "interplanetary_flight_to": {
                    "type": "string"
                },
                "is_primary": {
                    "type": "boolean"
                },
                "payload_mass_kg": {
                    "type": "number"
                },
                "propellant_kg": {
                    "type": "number"
                },
                "quantity": {
                    "type": "integer"
                },
                "route_id": {
                    "type": "integer"
                },
                "segment_dry_mass_kg": {
                    "type": "number"
                },
                "segment_isp_sec": {
                    "type": "number"
                },
                "segment_order": {
                    "type": "integer"
                },
                "stored_delta_v_ms": {
                    "description": "Сохранённые при формировании заявки (лаб. 2 / FormRequest)",
                    "type": "number"
                },
                "stored_fuel_mass_kg": {
                    "type": "number"
                },
                "to_orbit_radius_km": {
                    "type": "number"
                },
                "video_key": {
                    "type": "string"
                },
                "video_url": {
                    "type": "string"
                }
            }
        },
...
        "handler.ModerateInterplanetaryFlightRequestBody": {
            "type": "object",
            "required": [
                "action"
            ],
            "properties": {
                "action": {
                    "description": "\"complete\" | \"reject\"",
                    "type": "string"
                }
            }
        },
...
        "handler.UpdateInterplanetaryFlightRequestBody": {
            "type": "object",
            "properties": {
                "engine_isp_sec": {
                    "type": "number"
                },
                "spacecraft_dry_mass_kg": {
                    "type": "number"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}